Move `MockNestedHitObject` to own class
